@@ -1,44 +1,43 @@
 {
-  "name": "display-notification",
-  "version": "2.0.0",
-  "description": "Display a notification (macOS 10.9+)",
-  "license": "MIT",
-  "repository": "sindresorhus/display-notification",
-  "author": {
-    "name": "Sindre Sorhus",
-    "email": "sindresorhus@gmail.com",
-    "url": "sindresorhus.com"
-  },
-  "engines": {
-    "node": ">=4"
-  },
-  "scripts": {
-    "test": "xo && ava"
-  },
-  "files": [
-    "index.js"
-  ],
-  "keywords": [
-    "display",
-    "notification",
-    "notify",
-    "show",
-    "macos",
-    "osx",
-    "mac",
-    "applescript",
-    "center",
-    "notifier"
-  ],
-  "dependencies": {
-    "escape-string-applescript": "^1.0.0",
-    "run-applescript": "^3.0.0"
-  },
-  "devDependencies": {
-    "ava": "*",
-    "xo": "*"
-  },
-  "xo": {
-    "esnext": true
-  }
+	"name": "display-notification",
+	"version": "2.0.0",
+	"description": "Display a notification (macOS)",
+	"license": "MIT",
+	"repository": "sindresorhus/display-notification",
+	"funding": "https://github.com/sponsors/sindresorhus",
+	"author": {
+		"name": "Sindre Sorhus",
+		"email": "sindresorhus@gmail.com",
+		"url": "https://sindresorhus.com"
+	},
+	"type": "module",
+	"exports": "./index.js",
+	"engines": {
+		"node": ">=12"
+	},
+	"scripts": {
+		"test": "xo && ava"
+	},
+	"files": [
+		"index.js"
+	],
+	"keywords": [
+		"display",
+		"notification",
+		"notify",
+		"show",
+		"macos",
+		"mac",
+		"applescript",
+		"center",
+		"notifier"
+	],
+	"dependencies": {
+		"escape-string-applescript": "^3.0.0",
+		"run-applescript": "^5.0.0"
+	},
+	"devDependencies": {
+		"ava": "^3.15.0",
+		"xo": "^0.39.1"
+	}
 }
